ui: verifing the componentes in login screen, unit test
It's verifing the fields and button are used to login.
